LayoutTests/imported/w3c/web-platform-tests/html/dom/render-blocking/parser-inserted-defer-script-expected.txt

Some text

PASS Rendering is blocked before render-blocking resources are loaded
PASS Parser-inserted render-blocking defer script is evaluated
